data_analysis/requirements.txt
jupyterlab
ipython
pandas
matplotlib
seaborn
requests
python-dotenv
google-cloud-bigquery
db-dtypes
scikit-learn
pytorch-tabnet
pymongo
python-dotenv